backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.2.3
requests==2.31.0
supabase>=2.10.0
google-generativeai==0.3.0
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
